+++ esibd/devices/__init__.py
@@ -7,7 +7,6 @@ from .mips.mips import MIPS  # noqa: F401
 from .ni9263.ni9263 import NI9263  # noqa: F401
 from .omnicontrol.omnicontrol import OMNICONTROL  # noqa: F401
 from .pico.pico import PICO  # noqa: F401
-from .pressure.pressure import Pressure  # noqa: F401
 from .rbd.rbd import RBD  # noqa: F401
 from .rspd3303c.rspd3303c import RSPD3303C  # noqa: F401
 from .temperature.temperature import Temperature  # noqa: F401